pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pixshift"
version = "1.0.0"
description = "⚡ PixShift - 高效图片格式批量转换工具"
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
authors = [
    {name = "PixShift Team"},
]
keywords = ["image", "converter", "heic", "png", "batch", "format"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
]

dependencies = [
    "Pillow>=9.0",
    "pillow-heif>=0.10.0",
    "click>=8.0",
    "rich>=12.0",
    "PyMuPDF>=1.23.0",
]

[project.optional-dependencies]
avif = ["pillow-avif-plugin"]
all = ["pillow-avif-plugin"]
dev = ["pytest>=7.0", "pytest-cov>=4.0", "build>=1.2.0", "twine>=5.0.0"]

[project.scripts]
pixshift = "pixshift.cli:main"

[project.urls]
Homepage = "https://github.com/pixshift/pixshift"
Repository = "https://github.com/pixshift/pixshift"
Issues = "https://github.com/pixshift/pixshift/issues"
Documentation = "https://github.com/pixshift/pixshift#readme"

[tool.setuptools.packages.find]
include = ["pixshift*"]
